Moved CfgVehicles to its own file
